Gijs dev

+++ src/main/java/business/representation/RequestListRepresentation.java
@@ -17,7 +17,7 @@ public class RequestListRepresentation {
     private String title;
 
     private String approvalVote;
-    private Integer numberOfApprovalVotes;
+    private Long numberOfApprovalVotes;
     
     public RequestListRepresentation() {
 
@@ -104,11 +104,11 @@ public class RequestListRepresentation {
         this.approvalVote = approvalVote;
     }
 
-    public Integer getNumberOfApprovalVotes() {
+    public Long getNumberOfApprovalVotes() {
         return numberOfApprovalVotes;
     }
 
-    public void setNumberOfApprovalVotes(Integer numberOfApprovalVotes) {
+    public void setNumberOfApprovalVotes(Long numberOfApprovalVotes) {
         this.numberOfApprovalVotes = numberOfApprovalVotes;
     }
